requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
pytest 
httpx
python-dotenv